Fix weird numpy double type bug

+++ simple_diffusion_model/diffusion_wrapper.py
@@ -6,7 +6,7 @@ import numpy as np
 from torch.nn import Module
 
 def beta_schedule(timesteps):
-    return np.linspace(1e-4, 0.02, timesteps)
+    return np.linspace(1e-4, 0.02, timesteps).astype('float32')
 
 class DiffusionWrapper(nn.Module):
     def __init__(self, net, input_shape, timesteps=1000):
